Change Log
fix(handleMedia): display the entry status instead of unknown
refactor(utils): move some of the utils into a single file for better organization
