--- src/Utils/file.ts
@@ -1,3 +1,0 @@
-export function removeExtension(fileName: string, extension = '.ts') {
-  return fileName.slice(0, fileName.indexOf(extension))
-}
