+++ src/Utils/index.ts
@@ -6,14 +6,12 @@ export * from './Footer'
 export * from './GraphQLRequest'
 export * from './Logging'
 export * from './RSACryption'
-export * from './SeriesTitle'
 export * from './getOptions'
 export * from './types'
-export * from './file'
 export * from './registerCommands'
 export * from './registerEvents'
 export * from './registerComponents'
 export * from './getAnilistUser'
-export * from './normalize';
 export * from './getSubcommand';
 export * from './handleMediaData';
+export * from './common'
